refactor(shell): use kill instead of send_signal
send_signal send the message to the pgid first. If it failed, it sent
it to the pid. This was done because parent couldn't be sure child had
set pgid. Now that child notifies parent with SIGUSR1, parent can be
sure that pgid has been set and there's no more need of send_signal.